backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
langchain==0.3.7
langgraph==0.2.39
langchain-openai==0.2.10
openai==1.55.3
chromadb==0.5.5
sentence-transformers==3.2.1
pypdf==5.0.1
python-pptx==0.6.23
httpx==0.27.2